# Optimize trial decryption ordering