--- a/shadowsocks/tcp_test.go
+++ b/shadowsocks/tcp_test.go
@@ -15,13 +15,19 @@
 package shadowsocks
 
 import (
+	"errors"
+	"fmt"
+	"io"
 	"net"
 	"testing"
+	"time"
 
 	logging "github.com/op/go-logging"
+	"github.com/shadowsocks/go-shadowsocks2/shadowaead"
 )
 
-func BenchmarkTCPFindCipher(b *testing.B) {
+// Simulates receiving invalid TCP connection attempts on a server with 100 ciphers.
+func BenchmarkTCPFindCipherFail(b *testing.B) {
 	b.StopTimer()
 	b.ResetTimer()
 
@@ -52,5 +58,99 @@ func BenchmarkTCPFindCipher(b *testing.B) {
 		b.StartTimer()
 		findAccessKey(clientConn, cipherList)
 		b.StopTimer()
+	}
+}
+
+// Fake DuplexConn
+// 1-way pipe, representing the upstream flow as seen by the server.
+type conn struct {
+	clientAddr net.Addr
+	reader     io.ReadCloser
+	writer     io.WriteCloser
+}
+
+func (c *conn) Read(b []byte) (int, error) {
+	return c.reader.Read(b)
+}
+
+func (c *conn) Write(b []byte) (int, error) {
+	// Any downstream data is ignored.
+	return len(b), nil
+}
+
+func (c *conn) Close() error {
+	e1 := c.reader.Close()
+	e2 := c.writer.Close()
+	if e1 != nil {
+		return e1
+	}
+	return e2
+}
+
+func (c *conn) LocalAddr() net.Addr {
+	return nil
+}
+
+func (c *conn) RemoteAddr() net.Addr {
+	return c.clientAddr
+}
+
+func (c *conn) SetDeadline(t time.Time) error {
+	return errors.New("SetDeadline is not supported")
+}
+
+func (c *conn) SetReadDeadline(t time.Time) error {
+	return errors.New("SetDeadline is not supported")
+}
+
+func (c *conn) SetWriteDeadline(t time.Time) error {
+	return errors.New("SetDeadline is not supported")
+}
+
+func (c *conn) CloseRead() error {
+	return c.reader.Close()
+}
+
+func (c *conn) CloseWrite() error {
+	return nil
+}
+
+// Simulates receiving valid TCP connection attempts from 100 different users,
+// each with their own cipher and their own IP address.
+func BenchmarkTCPFindCipherRepeat(b *testing.B) {
+	b.StopTimer()
+	b.ResetTimer()
+
+	logging.SetLevel(logging.INFO, "")
+
+	const numCiphers = 100 // Must be <256
+	cipherList, err := MakeTestCiphers(numCiphers)
+	if err != nil {
+		b.Fatal(err)
+	}
+	cipherEntries := [numCiphers]*CipherEntry{}
+	for cipherNumber, element := range cipherList.SafeSnapshot(nil) {
+		cipherEntries[cipherNumber] = element.Value.(*CipherEntry)
+	}
+	for n := 0; n < b.N; n++ {
+		cipherNumber := n % numCiphers
+		reader, writer := io.Pipe()
+		addr := &net.TCPAddr{IP: net.ParseIP(fmt.Sprintf("192.0.2.%d", cipherNumber)), Port: 54321}
+		c := conn{addr, reader, writer}
+		go func() {
+			cipher := cipherEntries[cipherNumber].Cipher
+			salt := make([]byte, cipher.SaltSize())
+			writer.Write(salt)
+			encrypter, _ := cipher.Encrypter(salt)
+			ssWriter := shadowaead.NewWriter(writer, encrypter)
+			ssWriter.Write(MakeTestPayload(50))
+		}()
+		b.StartTimer()
+		_, _, err := findAccessKey(&c, cipherList)
+		b.StopTimer()
+		if err != nil {
+			b.Error(err)
+		}
+		c.Close()
 	}
 }