Ciphers last seen from the client's IP are tried first.
This is mostly relevant to servers that have many
users on separate keys, with simultaneous high-throughput
UDP flows.

Benchmark results before this change (using added benchmarks):
```
BenchmarkTCPFindCipherFail-40      	    1000	   1216804 ns/op
BenchmarkTCPFindCipherRepeat-40    	    2000	   1347786 ns/op
BenchmarkUDPUnpackFail-40          	    2000	    988158 ns/op
BenchmarkUDPUnpackRepeat-40        	    3000	   1014964 ns/op
BenchmarkUDPUnpackSharedKey-40     	  200000	     11505 ns/op
```

After this change:
```
BenchmarkTCPFindCipherFail-40      	    1000	   1499293 ns/op
BenchmarkTCPFindCipherRepeat-40    	   20000	     61834 ns/op
BenchmarkUDPUnpackFail-40          	    2000	    993359 ns/op
BenchmarkUDPUnpackRepeat-40        	  100000	     14883 ns/op
BenchmarkUDPUnpackSharedKey-40     	  200000	     10659 ns/op
```
I believe these benchmarks show no significant change in cases where all ciphers
must be tried, but a significant acceleration when each cipher is used
repeatedly by a different client IP address.